HW1 fix2

+++ Moiseev_Georgii/HW1/main.cpp
@@ -3,13 +3,14 @@
 #include <algorithm>
 #include <vector>
 #include <assert.h>
+#include <cstdint>
 #include "numbers.dat" 
 
 //table of primes on [0, n]
 class PrimeTable
 {
 public:
-    PrimeTable(int _n): n(_n)
+    PrimeTable(int32_t _n): n(_n)
     {
         assert(n > 1);
 
@@ -18,9 +19,9 @@ public:
         table[0] = false;
         table[1] = false;
 
-        for (int i = 2; i * i <= n; i++)
+        for (int32_t i = 2; i * i <= n; i++)
             if (table[i])
-                for (int j = i * i; j <= n; j += i)
+                for (int32_t j = i * i; j <= n; j += i)
                     table[j] = false;
     }
 
@@ -31,12 +32,12 @@ public:
 
 private:
     std::vector<bool> table;
-    int n;
+    int32_t n;
 
 };
 
 
-int primeCountArray(int lowerBound, int upperBound, int maxRightBound)
+int32_t primeCountArray(int32_t lowerBound, int32_t upperBound, int32_t maxRightBound)
 {
     static PrimeTable primeTable(maxRightBound);
 
@@ -56,7 +57,7 @@ int primeCountArray(int lowerBound, int upperBound, int maxRightBound)
         return 0;
 
     //now we know that upperBound exists in array
-    int ans = 0;
+    int32_t ans = 0;
     for (; (*leftIter) <= upperBound && leftIter < Data + Size; leftIter++)
         ans += primeTable[(*leftIter)] ? 1 : 0;
 
@@ -65,15 +66,15 @@ int primeCountArray(int lowerBound, int upperBound, int maxRightBound)
 
 int main(int argc, char* argv[])
 {
-    const int maxRightBound = 1e5;
+    const int32_t maxRightBound = 1e5;
     
     //incorrect input
     if (argc == 1 || argc % 2 == 0)
         return -1;
 
-    for (int i = 1; i < argc; ++i)
+    for (size_t i = 1; i < argc; ++i)
     {
-        int lowerBound, upperBound;
+        int32_t lowerBound, upperBound;
         //incorrect input
         if (sscanf(argv[i], "%d", &lowerBound) != 1)
             return -1;
